feat: replace sort and sortOrder to filterSort and sorting only works for search filter
