+++ README.md
@@ -100,6 +100,7 @@ export default () => (
 | disabled | whether disabled select | bool | false |
 | filterOption | whether filter options by input value. default filter by option's optionFilterProp prop's value | bool | true/Function(inputValue:string, option:Option) |
 | optionFilterProp | which prop value of option will be used for filter if filterOption is true | String | 'value' |
+| filterSort | Sort function for search options sorting, see [Array.sort](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort)'s compareFunction. | Function(optionA:Option, optionB: Option) | - |
 | optionLabelProp | render option value or option children as content of select | String: 'value'/'children' | 'value' |
 | defaultValue | initial selected option(s) | String/Array<String> | - |
 | value | current selected option(s) | String/Array<String>/{key:String, label:React.Node}/Array<{key, label}> | - |
@@ -127,8 +128,6 @@ export default () => (
 | loading | show loading icon in arrow | Boolean | false |
 | virtual | Disable virtual scroll | Boolean | true |
 | direction | direction of dropdown | 'ltr' \| 'rtl' | 'ltr' |
-| sorter | Sort function for local sort, see [Array.sort](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort)'s compareFunction. | Function(optionA:Option, optionB: Option) | - |
-| sortOrder | Order of sorted values: 'ascend' 'descend'  | string | 'ascend' |
 
 ### Methods
